1. SPI, UART, HAL and IMU/BMP sensors are now hiding the implementation through the pvMalloc and using PIMPL idiom 2. STM32.c hides HAL internals 3. removed uninon for altutide and added afunction float2bytes to deal with the conversions in (utils/common.h) 4. removed data_logger.h: not needed anymore since I moved al lthe functions into the sensor files 5. removed forward_declarations.h : not needed because of 1) 6. renamed: stm32f4xx_hal_uart_io.h -> AvionicsSoftware-AtollicProject/Inc/UART.h and corresponsing impl file 7. renamed: hardwareDefs.h -> AvionicsSoftware-AtollicProject/Inc/hardware_definitions.h 8. restructured main.c 9. replaced vTask creation with osThread creation in main / the same si for startup 10. Made a lot of refactoringand renaming of the files/variables and functions
This commit mainly:
1. Incapsulated many internal properties such as queues and driver handles via Memory Pool feature offered by FreeRTOS.
2. Change the tasks into threads.

The rest of the changes are just corresponding changes to be able to build the project and also some minor polishings/optimizations/improvements
im logic of the code.

+++ AvionicsSoftware-AtollicProject/Inc/flash.h
@@ -13,8 +13,9 @@
 #ifndef FLASH_H
 #define FLASH_H
 
-
+#include <stdio.h>
 #include <inttypes.h>
+
 /**
  * So far flash.h must depend on SPI.h because it initializes SPI_HandleTypeDef
  * @todo: maybe in the future we can use a pointer instead then?
@@ -183,6 +184,6 @@ uint8_t flash_get_status_register(Flash p_flash);
  * The address value (32 bits)
  * @see https://github.com/UMSATS/Avionics-2019/
  */
-uint32_t flash_scan(Flash p_flash);
+size_t flash_scan(Flash p_flash);
 
 #endif // FLASH_H
